requirements.txt
requests
pandas
pyarrow
duckdb
psycopg2-binary
qdrant-client
python-dotenv
jupyterlab
streamlit
